Refactored Gesture Interaction classes
