+++ lib/src/charts/painters/cartesian/cartesian_chart_painter.dart
@@ -8,123 +8,6 @@ import 'package:chart_it/src/charts/painters/cartesian/cartesian_painter.dart';
 import 'package:chart_it/src/charts/painters/text/chart_text_painter.dart';
 import 'package:flutter/material.dart';
 
-mixin InteractionDispatcher {
-  Offset? _latestDoubleTapOffset;
-  Offset? _latestPanOffset;
-
-  @protected
-  void onInteraction(
-      ChartInteractionType interactionType, Offset localPosition);
-
-  void onTapUp(Offset localPosition) {
-    onInteraction(ChartInteractionType.tap, localPosition);
-  }
-
-  void onDoubleTapDown(Offset localPosition) {
-    _latestDoubleTapOffset = localPosition;
-  }
-
-  void onDoubleTapCancel() {
-    _latestDoubleTapOffset = null;
-  }
-
-  void onDoubleTap() {
-    final latestDoubleTapOffset = _latestDoubleTapOffset;
-    if (latestDoubleTapOffset != null) {
-      onInteraction(ChartInteractionType.doubleTap, latestDoubleTapOffset);
-    }
-  }
-
-  void onPanStart(Offset localPosition) {
-    _latestPanOffset = localPosition;
-    onInteraction(ChartInteractionType.drag, localPosition);
-  }
-
-  void onPanUpdate(Offset localPosition) {
-    _latestPanOffset = localPosition;
-    onInteraction(ChartInteractionType.drag, localPosition);
-  }
-
-  void onPanCancel() {
-    _latestPanOffset = null;
-  }
-
-  void onPanEnd() {
-    final latestPanOffset = _latestPanOffset;
-    if (latestPanOffset != null) {
-      onInteraction(ChartInteractionType.dragEnd, latestPanOffset);
-    }
-    _latestPanOffset = null;
-  }
-}
-
-enum ChartInteractionType {
-  tap,
-  doubleTap,
-  dragStart,
-  drag,
-  dragEnd;
-}
-
-abstract class ChartInteractionResult {
-  final Offset? localPosition;
-  final ChartInteractionType interactionType;
-
-  ChartInteractionResult({
-    required this.localPosition,
-    required this.interactionType,
-  });
-}
-
-abstract class ChartInteractionConfig<T extends ChartInteractionResult> {
-  final bool isEnabled;
-  final void Function(T interactionResult)? onRawInteraction;
-  final void Function(T interactionResult)? onTap;
-  final void Function(T interactionResult)? onDoubleTap;
-  final void Function(T interactionResult)? onDragStart;
-  final void Function(T interactionResult)? onDrag;
-  final void Function(T interactionResult)? onDragEnd;
-
-  const ChartInteractionConfig({
-    this.onRawInteraction,
-    required this.onTap,
-    required this.onDoubleTap,
-    required this.onDragStart,
-    required this.onDrag,
-    required this.onDragEnd,
-    required this.isEnabled,
-  });
-
-  void onInteraction(T interactionResult) {
-    switch (interactionResult.interactionType) {
-      case ChartInteractionType.tap:
-        onTap?.call(interactionResult);
-        break;
-      case ChartInteractionType.doubleTap:
-        onDoubleTap?.call(interactionResult);
-        break;
-      case ChartInteractionType.drag:
-        onDragStart?.call(interactionResult);
-        break;
-      case ChartInteractionType.dragStart:
-        onDrag?.call(interactionResult);
-        break;
-      case ChartInteractionType.dragEnd:
-        onDragEnd?.call(interactionResult);
-        break;
-    }
-  }
-
-  bool get shouldHitTest =>
-      isEnabled &&
-      (onRawInteraction != null ||
-          onTap != null ||
-          onDoubleTap != null ||
-          onDragStart != null ||
-          onDragEnd != null ||
-          onDrag != null);
-}
-
 class CartesianChartPainter {
   late Rect graphPolygon;
   late double graphHeight;
