+++ lib/src/interactions/interactions.dart
@@ -1,0 +1,3 @@
+export 'data/chart_interaction_config.dart';
+export 'data/chart_interactions.dart';
+export 'hit_test/interaction_dispatcher.dart';
